Setup the Button - Did some refactoring - Why is my renaming function not working 😔
